+++ src/modules/feature/board/board/BoardHeader.unit.ts
@@ -12,7 +12,6 @@ import {
 } from "@@/tests/test-utils/setup";
 import { useBoardFocusHandler, useBoardPermissions } from "@data-board";
 import {
-	BoardAnyTitleInput,
 	BoardMenuActionCopy,
 	BoardMenuActionDelete,
 	BoardMenuActionEdit,
@@ -23,6 +22,7 @@ import {
 import { useCourseBoardEditMode } from "@util-board";
 import { shallowMount } from "@vue/test-utils";
 import { computed } from "vue";
+import BoardAnyTitleInput from "../shared/BoardAnyTitleInput.vue";
 import BoardHeader from "./BoardHeader.vue";
 
 jest.mock("@data-board");
