+++ src/modules/feature/board/index.ts
@@ -1,3 +1,5 @@
 import Board from "./board/Board.vue";
+import BoardColumnInteractionHandler from "./board/BoardColumnInteractionHandler.vue";
+import BoardAnyTitleInput from "./shared/BoardAnyTitleInput.vue";
 
-export { Board };
+export { Board, BoardColumnInteractionHandler, BoardAnyTitleInput };
